YT-24913: Split memory tracker commit_hash:af573daeadda0ee67ff6c0a595b6dc3ad6314c51
