+++ library/cpp/yt/memory/chunked_output_stream.h
@@ -22,7 +22,6 @@ class TChunkedOutputStream
 public:
     explicit TChunkedOutputStream(
         TRefCountedTypeCookie tagCookie = GetRefCountedTypeCookie<TDefaultChunkedOutputStreamTag>(),
-        IMemoryUsageTrackerPtr memoryUsageTracker = GetNullMemoryUsageTracker(),
         size_t initialReserveSize = 4_KB,
         size_t maxReserveSize = 64_KB);
 
@@ -48,9 +47,6 @@ public:
     void Advance(size_t size);
 
 private:
-    const IMemoryUsageTrackerPtr MemoryUsageTracker_;
-
-    TMemoryUsageTrackerGuard CurrentChunkMemoryUsageGuard_;
     size_t MaxReserveSize_;
     size_t CurrentReserveSize_;
 
@@ -59,10 +55,7 @@ private:
     TBlob CurrentChunk_;
     std::vector<TSharedRef> FinishedChunks_;
 
-
     void ReserveNewChunk(size_t spaceRequired);
-
-    void UpdateCurrentChunkMemoryUsage();
 
     void DoWrite(const void* buf, size_t len) override;
     size_t DoNext(void** ptr) override;
