+++ library/cpp/yt/memory/public.h
@@ -12,8 +12,7 @@ constexpr size_t CacheLineSize = 64;
 class TChunkedMemoryPool;
 
 DECLARE_REFCOUNTED_STRUCT(IMemoryChunkProvider)
-DECLARE_REFCOUNTED_STRUCT(IMemoryUsageTracker)
-DECLARE_REFCOUNTED_STRUCT(IReservingMemoryUsageTracker)
+DECLARE_REFCOUNTED_STRUCT(ISimpleMemoryUsageTracker)
 DECLARE_REFCOUNTED_STRUCT(TSharedRangeHolder)
 
 using TMemoryTag = ui32;
