+++ yt/yt/client/driver/driver.h
@@ -13,6 +13,7 @@
 #include <yt/yt/core/concurrency/async_stream.h>
 
 #include <yt/yt/core/misc/error.h>
+#include <yt/yt/core/misc/memory_usage_tracker.h>
 
 #include <yt/yt/core/net/address.h>
 
@@ -22,8 +23,6 @@
 #include <yt/yt/core/yson/writer.h>
 
 #include <yt/yt/core/ytree/public.h>
-
-#include <library/cpp/yt/memory/memory_usage_tracker.h>
 
 namespace NYT::NDriver {
