+++ yt/yt/core/bus/tcp/client.h
@@ -4,7 +4,7 @@
 
 #include "packet.h"
 
-#include <library/cpp/yt/memory/memory_usage_tracker.h>
+#include <yt/yt/core/misc/memory_usage_tracker.h>
 
 namespace NYT::NBus {
