+++ yt/yt/core/bus/tcp/connection.h
@@ -24,7 +24,7 @@
 
 #include <yt/yt/core/concurrency/pollable_detail.h>
 
-#include <library/cpp/yt/memory/memory_usage_tracker.h>
+#include <yt/yt/core/misc/memory_usage_tracker.h>
 
 #include <library/cpp/yt/threading/atomic_object.h>
 #include <library/cpp/yt/threading/spin_lock.h>
