+++ yt/yt/core/https/server.h
@@ -8,7 +8,7 @@
 
 #include <yt/yt/core/http/public.h>
 
-#include <library/cpp/yt/memory/memory_usage_tracker.h>
+#include <yt/yt/core/misc/memory_usage_tracker.h>
 
 namespace NYT::NHttps {
