+++ yt/yt/core/misc/concurrent_cache.h
@@ -3,8 +3,7 @@
 #include "public.h"
 #include "atomic_ptr.h"
 #include "lock_free_hash_table.h"
-
-#include <library/cpp/yt/memory/memory_usage_tracker.h>
+#include "memory_usage_tracker.h"
 
 namespace NYT {
