+++ yt/yt/core/misc/public.h
@@ -205,4 +205,9 @@ DEFINE_ENUM(EProcessErrorCode,
 
 ////////////////////////////////////////////////////////////////////////////////
 
+DECLARE_REFCOUNTED_STRUCT(IMemoryUsageTracker)
+DECLARE_REFCOUNTED_STRUCT(IReservingMemoryUsageTracker)
+
+////////////////////////////////////////////////////////////////////////////////
+
 } // namespace NYT
