+++ yt/yt/core/rpc/channel.h
@@ -10,6 +10,7 @@
 #include <yt/yt/core/bus/client.h>
 
 #include <yt/yt/core/misc/error.h>
+#include <yt/yt/core/misc/memory_usage_tracker.h>
 
 #include <yt/yt/core/compression/public.h>
