+++ yt/yt/core/rpc/server_detail.h
@@ -7,9 +7,9 @@
 
 #include <yt/yt/core/logging/log.h>
 
-#include <yt/yt_proto/yt/core/rpc/proto/rpc.pb.h>
+#include <yt/yt/core/misc/memory_usage_tracker.h>
 
-#include <library/cpp/yt/memory/memory_usage_tracker.h>
+#include <yt/yt_proto/yt/core/rpc/proto/rpc.pb.h>
 
 #include <library/cpp/yt/threading/rw_spin_lock.h>
 #include <library/cpp/yt/threading/spin_lock.h>
