+++ yt/yt/core/ytree/ypath_detail.cpp
@@ -6,6 +6,8 @@
 #include "system_attribute_provider.h"
 #include "ypath_client.h"
 
+#include <yt/yt/core/misc/memory_usage_tracker.h>
+
 #include <yt/yt/core/yson/attribute_consumer.h>
 
 #include <yt/yt/core/ypath/tokenizer.h>
